disable logout btn when not logged in;
