--- client/components/Auth/Auth.jsx
@@ -1,12 +1,0 @@
-import React from 'react'
-
-import LogOut from './LogOut'
-const Auth = () => {
-  return (
-    <div className='main-container'>
-      <LogOut />
-    </div>
-  )
-}
-
-export default Auth
